+++ datastore/get_persistence_infos.go
@@ -51,7 +51,7 @@ func (commonProtocol *CommonProtocol) getPersistenceInfos(err error, packet nex.
 		}
 
 		// * The owner of an object can always view their objects, but normal users cannot
-		if metaInfo.Status != types.UInt8(datastore_constants.DataStatusNone) && metaInfo.OwnerID != connection.PID() {
+		if metaInfo.Status != datastore_constants.DataStatusNone && metaInfo.OwnerID != connection.PID() {
 			pPersistenceInfo = append(pPersistenceInfo, invalidPersistenceInfo)
 			pResults = append(pResults, types.NewQResultError(nex.ResultCodes.DataStore.NotFound))
 			continue
