+++ datastore/get_specific_meta.go
@@ -20,7 +20,7 @@ func (commonProtocol *CommonProtocol) getSpecificMeta(err error, packet nex.Pack
 	connection := packet.Sender()
 	endpoint := connection.Endpoint()
 
-	if len(param.DataIDs) > int(datastore_constants.BatchProcessingCapacity) {
+	if len(param.DataIDs) > datastore_constants.BatchProcessingCapacity {
 		return nil, nex.NewError(nex.ResultCodes.DataStore.InvalidArgument, "change_error")
 	}
 
@@ -45,8 +45,8 @@ func (commonProtocol *CommonProtocol) getSpecificMeta(err error, packet nex.Pack
 		}
 
 		// * The owner of an object can always view their objects, but normal users cannot
-		if metaInfo.Status != types.UInt8(datastore_constants.DataStatusNone) && metaInfo.OwnerID != connection.PID() {
-			if metaInfo.Status == types.UInt8(datastore_constants.DataStatusPending) {
+		if metaInfo.Status != datastore_constants.DataStatusNone && metaInfo.OwnerID != connection.PID() {
+			if metaInfo.Status == datastore_constants.DataStatusPending {
 				return nil, nex.NewError(nex.ResultCodes.DataStore.UnderReviewing, "change_error")
 			}
